Rename submodule to code
