+++ src/js/api.ts
@@ -51,7 +51,7 @@ API.runPythonInternal = function (code: string): any {
 
 let runPythonPositionalGlobalsDeprecationWarned = false;
 /**
- * Runs a string of Python code from JavaScript, using :any:`pyodide.eval_code`
+ * Runs a string of Python code from JavaScript, using :any:`pyodide.code.eval_code`
  * to evaluate the code. If the last statement in the Python code is an
  * expression (and the code doesn't end with a semicolon), the value of the
  * expression is returned.
@@ -68,7 +68,7 @@ let runPythonPositionalGlobalsDeprecationWarned = false;
  * @param options.globals An optional Python dictionary to use as the globals.
  *        Defaults to :any:`pyodide.globals`.
  * @returns The result of the Python code translated to JavaScript. See the
- *          documentation for :any:`pyodide.eval_code` for more info.
+ *          documentation for :any:`pyodide.code.eval_code` for more info.
  */
 export function runPython(
   code: string,
@@ -86,7 +86,7 @@ export function runPython(
   if (!options.globals) {
     options.globals = API.globals;
   }
-  return API.pyodide_eval.eval_code(code, options.globals);
+  return API.pyodide_code.eval_code(code, options.globals);
 }
 API.runPython = runPython;
 
@@ -116,7 +116,7 @@ export async function loadPackagesFromImports(
   messageCallback?: (msg: string) => void,
   errorCallback?: (err: string) => void
 ) {
-  let pyimports = API.pyodide_eval.find_imports(code);
+  let pyimports = API.pyodide_code.find_imports(code);
   let imports;
   try {
     imports = pyimports.toJs();
@@ -141,7 +141,7 @@ export async function loadPackagesFromImports(
 
 /**
  * Run a Python code string with top level await using
- * :any:`pyodide.eval_code_async` to evaluate the code. Returns a promise which
+ * :any:`pyodide.code.eval_code_async` to evaluate the code. Returns a promise which
  * resolves when execution completes. If the last statement in the Python code
  * is an expression (and the code doesn't end with a semicolon), the returned
  * promise will resolve to the value of this expression.
@@ -196,7 +196,7 @@ export async function runPythonAsync(
   if (!options.globals) {
     options.globals = API.globals;
   }
-  return await API.pyodide_eval.eval_code_async(code, options.globals);
+  return await API.pyodide_code.eval_code_async(code, options.globals);
 }
 API.runPythonAsync = runPythonAsync;
